Add custom exception, change hidden numbers behaviour

+++ src/main/java/pl/sda/evaluator/AsciiCharacterEvaluationStrategy.java
@@ -1,11 +1,12 @@
 package pl.sda.evaluator;
 
 public class AsciiCharacterEvaluationStrategy implements CharacterEvaluationStrategy {
+
     @Override
     public int evaluate(Character character) {
         if(character == null) {
             return 0;
         }
-        return character;
+        return (int) character;
     }
 }
